+++ x/group/internal/math/dec.go
@@ -1,6 +1,9 @@
+// Package math provides helper functions for doing mathematical calculations and parsing for the ecocredit module.
 package math
 
 import (
+	"fmt"
+
 	"github.com/cockroachdb/apd/v2"
 
 	"github.com/cosmos/cosmos-sdk/types/errors"
@@ -16,14 +19,10 @@ type Dec struct {
 	dec apd.Decimal
 }
 
-const mathCodespace = "math"
-
-var ErrInvalidDecString = errors.Register(mathCodespace, 1, "invalid decimal string")
-
 func NewDecFromString(s string) (Dec, error) {
 	d, _, err := apd.NewFromString(s)
 	if err != nil {
-		return Dec{}, ErrInvalidDecString.Wrap(err.Error())
+		return Dec{}, errors.ErrInvalidDecString.Wrap(err.Error())
 	}
 	return Dec{*d}, nil
 }
@@ -64,4 +63,24 @@ func (x Dec) IsEqual(y Dec) bool {
 
 func (x Dec) IsNegative() bool {
 	return x.dec.Negative && !x.dec.IsZero()
+}
+
+// Add adds x and y
+func Add(x Dec, y Dec) (Dec, error) {
+	return x.Add(y)
+}
+
+// SubNonNegative subtracts the value of y from x and returns the result with
+// arbitrary precision. Returns an error if the result is negative.
+func SubNonNegative(x Dec, y Dec) (Dec, error) {
+	z, err := x.Sub(y)
+	if err != nil {
+		return Dec{}, err
+	}
+
+	if z.IsNegative() {
+		return z, fmt.Errorf("result negative during non-negative subtraction")
+	}
+
+	return z, nil
 }
